Update for photostand script: don't explicitly test for a connected barcode scanner, but do test for operation outside an X display. Use the right comment syntax for .sh in photostand.sh, and don't both trying to remove old files if they aren't there.

+++ extras/scripts/photostand.sh
@@ -41,10 +41,10 @@ command -v wmctrl >/dev/null && wmctrl -a photostand
 
 . "$LIB_DIR"/photo-preamble.sh
 . "$LIB_DIR"/photo-functions.sh
-READ_BARCODE="$LIB_DIR"/read_barcode.py
+# READ_BARCODE="$LIB_DIR"/read_barcode.py
 
-rm uploads.log > /dev/null
-rm checkins.log > /dev/null
+[[ -f uploads.log ]] && rm uploads.log > /dev/null
+[[ -f checkins.log ]] && rm checkins.log > /dev/null
 
 killall_gvfs_volume_monitor
 
@@ -65,7 +65,7 @@ while true ; do
         sleep 5s
         continue
     fi
-    // BARCODE=`$READ_BARCODE "$DEV"`
+    # BARCODE=`$READ_BARCODE "$DEV"`
     read BARCODE
     echo Scanned $BARCODE
     CAR_NO=`echo $BARCODE | grep -e "^PWD" | sed -e "s/^PWD//"`
